+++ lndocs/lamin_sphinx/_authors.py
@@ -1,0 +1,6 @@
+authors = {
+    "falexwolf": ("Alex Wolf", "https://falexwolf.me"),
+    "sunnyosun": ("Sunny Sun", "https://github.com/sunnyosun"),
+    "Koncopd": ("Sergei Rybakov", "https://github.com/Koncopd"),
+    "Zethson": ("Lukas Heumos", "https://github.com/Zethson"),
+}
