+++ lndocs/lamin_sphinx/_nitpick_ignore.py
@@ -1,0 +1,24 @@
+nitpick_ignore = [
+    ("py:class", "pandas.core.frame.DataFrame"),
+    ("py:class", "sqlmodel.orm.session.Session"),
+    ("py:class", "numpy.ndarray"),
+    ("py:class", "datetime.datetime"),
+    ("py:class", "pydantic.main.BaseModel"),
+    ("py:class", "cloudpathlib.cloudpath.CloudPath"),
+    ("py:class", "anndata._core.anndata.AnnData"),
+    ("py:class", "enum.Enum"),
+    ("py:class", "pathlib.Path"),
+    ("py:class", "Model"),
+    ("py:class", "sqlmodel.main.SQLModel"),
+    ("py:class", "MetaData"),
+    ("py:class", "DictStrAny"),
+    ("py:class", "unicode"),
+    ("py:class", "typing.DictStrAny"),
+    ("py:class", "typing.unicode"),
+    ("py:data", "typing.Optional"),
+    ("py:data", "typing.Literal"),
+    ("py:data", "typing.Union"),
+    ("py:data", "typing.Any"),
+    ("py:obj", "bionty.Gene"),
+    ("py:obj", "bionty.Species"),
+]
